% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{getResiduals}
\alias{getResiduals}
\title{Regresses out covariates}
\usage{
getResiduals(expr, samplesheet, name, cov)
}
\arguments{
\item{expr}{A matrix with features on the rows and individual on the columns}

\item{samplesheet}{data.frame of meta data, columns corresponding to the rows
in expression}

\item{name}{names of the features}

\item{cov}{vector of strings corresponding to columns in the samplesheet that need
to be regressed out in the same order as expression. Rows in expression are
are columns in sample}
}
\description{
Regresses out covariates
}
\examples{
\dontrun{
library(gsQTL)
}
}
