% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePathwayExpression.R
\name{makePathwayExpression}
\alias{makePathwayExpression}
\title{Creates a gene set matrix from a gene expression matrix'}
\usage{
makePathwayExpression(
  pathwayInfo,
  expression,
  minGenes,
  minExplained,
  regressOut = F,
  samplesheet = NULL,
  cov = NULL,
  linear = TRUE,
  sigma = 10
)
}
\arguments{
\item{pathwayInfo}{A data.frame with two columns: pathname and symbol
where pathname represent the gene set/pathway and symbol represents the genes
belonging to that pathway}

\item{expression}{A matrix with features on the rows and individual on the columns}

\item{minGenes}{Minimal number of genes a gene set must consists of. This is
the minimal number after filtering the expression matrix based on features
that overlap with pathwayInfo}

\item{minExplained}{minimal requirement of variance explained of PC1}

\item{regressOut}{Boolean stating whether variables need to be regressed out}

\item{samplesheet}{data.frame of meta data, columns corresponding to the rows
in expression}

\item{cov}{vector of strings corresponding to columns in the samplesheet that need
to be regressed out in the same order as expression. Rows in expression are
are columns in sample}

\item{linear}{Boolean indicating whether linear or kernel PCA needs to be used}

\item{sigma}{inverse kernel width for the Radial Basis kernel}
}
\value{
Matrix of gene set by ID
}
\description{
Creates a gene set matrix from a gene expression matrix'
}
\examples{
\dontrun{
library(gsQTL)
gsQTLs
}
}
