% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLinearModel.R
\name{runLinearModel}
\alias{runLinearModel}
\title{Runs linear model}
\usage{
runLinearModel(
  expression,
  samplesheet,
  predictor,
  cov = NULL,
  padj,
  type = "linear"
)
}
\arguments{
\item{expression}{A matrix with features on the rows and individual on the columns}

\item{samplesheet}{data.frame of meta data, columns corresponding to the rows
in expression}

\item{predictor}{predictor variable for the linear model. Should be a column
in samplesheet}

\item{cov}{vector of strings corresponding to columns in the samplesheet that need
to be regressed out in the same order as expression. Rows in expression are
are columns in sample}

\item{padj}{type of multiple testing correction. See p.adjust}

\item{type}{string linear or LR for logistic regression}
}
\value{
Matrix of gene set by ID
}
\description{
Runs linear model
}
\examples{
\dontrun{
library(gsQTL)
}
}
